daemon/requirements.txt
# PropPilot AI — Python VPS Daemon
# Install: pip install -r requirements.txt

# Supabase
supabase==2.10.0
postgrest==0.18.0

# HTTP / WebSocket
httpx==0.27.2
websockets>=13.0
aiohttp>=3.10.11

# Data processing
pandas==2.2.2
numpy==1.26.4

# Scheduling
apscheduler==3.10.4

# AI / LLM  (Groq — FREE tier, no credit card)
openai==1.35.0

# Utilities
python-dotenv==1.0.1
structlog==24.4.0
tenacity==8.5.0        # retry logic
pytz==2024.1
